Add Constants for JOSE Identifiers (Headers, Parameters & Claims)
JOSE Identifiers are a known quantity, yet there are no constants that are available for consumer use, leaving them either to make their own or sprinkle hard-coded strings like `"typ"` throughout their code.

This commit adds constants that provide all the identifiers that are supported by this library. 

The first set is the `JOSEIdentifiers` which contains those header parameters intended for use in a JWE or JWS.

The second set is the `JWKIdentifiers` that contains the identifiers that are specific only to JWKs. 

The third set is the `JWTIdentifiers` which contains the registered claims intended for use in a JWT payload.

issue #433